services/tts/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
piper-tts==1.2.0
pydantic==2.10.4
aiofiles==24.1.0
